issue-33 Final adjustments to the module
Remove anything related to Chatbot or GuardDuty. Let's keep this module for just the Cloudtrail setup.
I think that leaves only one reference to modules/iam_service_role. If that's true, remove the module and put the resources in a top-level file. Having a module that is only used once feels unnecessary.